fixup! Create project context from header and authorize based on user context
