+++ backend/WebApi/Constants.cs
@@ -2,5 +2,5 @@ namespace LanguageForge.WebApi;
 
 public static class Constants
 {
-    public const string ProjectCodeHeader = "Project-Code";
+    public const string ProjectCodeHeader = "project-code";
 }
