+++ backend/WebApi/Program.cs
@@ -4,16 +4,13 @@ using LanguageForge.Api.Configuration;
 using LanguageForge.Api.Entities;
 using LanguageForge.WebApi;
 using LanguageForge.WebApi.Auth;
-using LanguageForge.WebApi.Validation;
 using Microsoft.OpenApi.Models;
 
 var builder = WebApplication.CreateBuilder(args);
 
 // Add services to the container.
 
-builder.Services.AddControllers()
-.AddMvcOptions(options => options.Filters.Add<RequireProjectCodeFilter>())
-.AddJsonOptions(opts =>
+builder.Services.AddControllers().AddJsonOptions(opts =>
 {
     var enumConverter = new JsonStringEnumConverter();
     opts.JsonSerializerOptions.Converters.Add(enumConverter);
